🚚 move indexer logger in infrastructure
