+++ marketplace-indexer/Cargo.toml
@@ -53,13 +53,6 @@ prost-types = "0.10.1"
 
 # Logs
 log = "0.4.17"
-slog = "2.7.0"
-slog-async = "2.7.0"
-slog-envlogger = "2.2.0"
-slog-json = "2.6.1"
-slog-scope = "4.4.0"
-slog-stdlog = "4.1.1"
-slog-term = "2.9.0"
 
 # Unit tests
 mockall = "0.11.2"
