+++ marketplace-indexer/src/main.rs
@@ -8,50 +8,13 @@ use dotenv::dotenv;
 use infrastructure::single_contract::SingleContract;
 use log::{error, info};
 use marketplace_domain::*;
-use marketplace_infrastructure::{database, event_webhook::EventWebHook, github, starknet};
-use slog::Logger;
+use marketplace_infrastructure::{database, event_webhook::EventWebHook, github, logger, starknet};
 use std::sync::Arc;
-
-fn channel_size() -> usize {
-	std::env::var("SLOG_CHANNEL_SIZE").unwrap_or_default().parse().unwrap_or(256)
-}
-
-fn get_root_logger() -> Logger {
-	use slog::{o, Drain, FnValue, Record};
-	use slog_async::Async;
-	use slog_json::Json;
-	use slog_term::{CompactFormat, TermDecorator};
-	use std::io::stdout;
-
-	let drain = match std::env::var("LOGS") {
-		Ok(logs) if &logs == "terminal" => Async::new(slog_envlogger::new(
-			CompactFormat::new(TermDecorator::new().stderr().build()).build().fuse(),
-		))
-		.chan_size(channel_size())
-		.build(),
-
-		_ => Async::new(slog_envlogger::new(
-			Json::new(stdout())
-				.add_default_keys()
-				.add_key_value(o!("location" => FnValue(move |record : &Record| {
-						format!("{}:{}:{}", record.file(), record.line(), record.column())
-					}),
-				))
-				.build()
-				.fuse(),
-		))
-		.chan_size(channel_size())
-		.build(),
-	};
-	slog_stdlog::init().unwrap();
-	Logger::root(drain.fuse(), o!("version" => env!("CARGO_PKG_VERSION")))
-}
 
 #[tokio::main]
 async fn main() {
 	dotenv().ok();
-	let global_logger_guard = slog_scope::set_global_logger(get_root_logger());
-	global_logger_guard.cancel_reset();
+	logger::new_global_logger();
 	github::Client::initialize();
 
 	let database = Arc::new(database::Client::new(database::init_pool()));
