+++ marketplace-infrastructure/Cargo.toml
@@ -51,6 +51,13 @@ crypto-bigint = { version = "0.4.8", features = ["serde"] }
 
 # Logs
 log = "0.4.17"
+slog = "2.7.0"
+slog-async = "2.7.0"
+slog-envlogger = "2.2.0"
+slog-json = "2.6.1"
+slog-scope = "4.4.0"
+slog-stdlog = "4.1.1"
+slog-term = "2.9.0"
 
 # Errors
 anyhow = "1.0.57"
